Add files via upload
The second .m file is to find the MS at different z. As show of the mainsequence.png file:.
I divide the Galaxies by stellar mass. And find the maximum points at each bins (I did this by the first .m file).  I set the value of the maximum points of each bins as 1 and take uniform scaling so that the red continuous line can somehow represent as MS.  
And the blue nonlinear curve is I use method of interpolation to find the actual main sequence at z=0.