Findthecut.m; quencut.png
We have no interest in the quiescent galaxies, so I manage to make a quench cut so that "active galaxies" can be well-defined. 
I complete this operation by the Findthecut.m, I set the slope of the cut line same as MS, and adjust its intercept to make the number of galaxies above the MS same as the galaxies below the MS.
As show in the quencut.png; The total number of galaxies above the MS is equal to the number between the MS and red cut line.