Add files via upload
This fig shows how to find the maximum points at each bins. The fig shows the distribution of SFR at fixed stellar mass, so that I can find the peak of the distribution which represents the maximum points. And at high stellar mass area, to find the peak I widen the width of bin because the samples are few.