Add files via upload
We are more interested in the Central Galaxies than Satellite Galaxies because the CG have more tight link to the property of the Halo it belongs to. As shown in the CG MS.png, I manage to find the Main Sequence of CG and compare it with the MS defined by all the active galaxies, and the result shows that the difference is little, especially at low redshift.
And the cgratio.png shows the fraction of central galaxies within active galaxies at fixed stellar mass. And the errorbar is from the Jackinife's resampple Method, to devide the universe into 8 cubic and calculate the ratio by ignore one of them.